backend_django/requirements.txt
#############################################
# VisionCare Backend Consolidated Requirements
#
# Generated by cleanup (merged duplicate files)
# Contains:
#  - Django REST stack
#  - Auth (SimpleJWT)
#  - DB utilities
#  - AI / Image processing & ONNX runtime
#  - PDF generation & supporting libs
#  - OpenAI client
#  - HTTP / async support libs
#############################################

# --- Web Framework & REST ---
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.8.0
dj-database-url==3.0.1
asgiref==3.9.1
sqlparse==0.5.3

# --- Auth & Security ---
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1

# --- Environment ---
python-dotenv==1.1.1

# --- Data Validation ---
pydantic==2.11.9
pydantic_core==2.33.2
annotated-types==0.7.0
typing-inspection==0.4.1
typing_extensions==4.15.0

# --- HTTP Clients / Async ---
# Use versions compatible with supabase==2.6.0 and openai==1.108.0
# supabase 2.6.0 requires httpx < 0.28; openai allows <1.
httpx==0.27.2
httpcore
anyio==4.10.0
sniffio==1.3.1
h11
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
distro==1.9.0
jiter==0.11.0

# --- AI / Image Processing ---
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
onnxruntime==1.20.0
tqdm==4.67.1
supabase==2.6.0

# --- PDF Generation ---
reportlab==4.4.3

# NOTE: If deploying to production with PostgreSQL, ensure psycopg2-binary
psycopg2-binary==2.9.10

# (End of file)
